multi-sdf

+++ dsdf/netowrk.py
@@ -46,7 +46,7 @@ class DSDFNetwork(nn.Module):
         # x: [B, 4]
 
         #import ipdb; ipdb.set_trace()
-        x = x.float()
+        x = x.double()
         x = self.encoder(x)
         x = x.double()
